% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_day_month.R
\name{first_day_month}
\alias{first_day_month}
\title{find first day of month}
\usage{
first_day_month(start_yr, end_yr)
}
\arguments{
\item{start_yr}{Numeric year to begin calendar}

\item{end_yr}{Numeric year to end calendar}
}
\value{
data.frame \code{monthly} in the years passed to function saved as
"data/monthly.RData".
}
\description{
{
first_day_month creates a data.frame with the first day of the month
to be used in the tidy data process.  The first day of the month is used as a
trade entry criteria in options.shiny and options.studies.
}
}
\examples{
monthly <- monthly(2010, 2020)

}

