% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_closed.R
\name{mrkt_closed}
\alias{mrkt_closed}
\title{market closed dates}
\usage{
mrkt_closed(file_name)
}
\arguments{
\item{file_name}{String of file to be read in.}
}
\value{
data.frame \code{market_closed} with the each day the market was closed saved
as "data/market_closed.RData".
}
\description{
{
mrkt_closed creates a data.frame with the dates the market was closed by reading
from .csv file passed to be used in the tidy data process.  Update the .csv prior to
 running.
}
}
\examples{
market_closed <- mrkt_closed("data/market_closed.csv")
}

