% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_short_put.R
\name{short_put}
\alias{short_put}
\title{Short put options strategy}
\usage{
short_put(progress.int, t)
}
\description{
{
short_put opens a short put trade that meet criteria chosen in the Shiny UI from shinyoptions
}
}
\examples{

}

