% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_strangle.R
\name{strangle}
\alias{strangle}
\title{Strangle options strategy}
\usage{
strangle(progress.int, t)
}
\description{
{
strangle opens a short strangle trade that meet criteria chosen in the Shiny UI
}
}
\examples{

}

