% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/third_friday.R
\name{third_friday}
\alias{third_friday}
\title{Find third friday of month}
\usage{
third_friday(start_yr, end_yr)
}
\arguments{
\item{start_yr}{Numeric year to begin third friday  calendar}

\item{end_yr}{Numeric year to end third friday calendar}
}
\value{
data.frame \code{monthly_exp_date} with the third friday of each month
when options expire.
}
\description{
{
third_friday creates a data.frame with options expiration dates to be used in
the tidy data process.  The third friday date identifies monthly expirations
compared to weeklys.
}
}
\examples{
third_friday(2010, 2015)
}

