% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_options.R
\name{tidy_options}
\alias{tidy_options}
\title{Tidy iVolatility options .csv files}
\usage{
tidy_options(ticker, opt_path, iv_file)
}
\arguments{
\item{opt_path}{Character string of path to iVolatility files

  When files are downloaded from iVolatility they are limited in size to 50Mb.
  Place all the files in the "opt_path" folder under your working directory.}

\item{iv_file}{Character string of volatility file for symbol chosen

  IVRank data from \href{http://www.cboe.com/micro/equityvix/introduction.aspx}{http://www.cboe.com/micro/equityvix/introduction.aspx}
  make sure to remove the disclaimer at the top of the file from this site.

  IV can also be calculated outside this process and used if the file format
  is the same.}

\item{symbol}{Character string for symbol you would like to process

  Value will be used to name the output .RData file}
}
\value{
\code{tidy_options} returns tidy options chain data with
 (rsi_14, iv_rank_252, and iv_rank_90) added.

 Output can then be used in \code{options.shiny} and \code{options.studies}
 for backtesting.
}
\description{
{
tidy_options produces a tidy dataset from downloaded iVolatility.com files.
Download data from \href{http://www.ivolatility.com/data_download.j}{http://www.ivolatility.com/data_download.j}

   \itemize{
      \item Individual contracts (Raw IV) data
      \item include greeks in file download
      \item place files in folder under the working directory
      \item ensure that market closed dates are current
    }
}
}
\examples{
tidy_options("XLE", "data/raw_files/options", "data/volatility/vx.xle.daily.prices.RData")
}

