% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_tooltip.R
\name{trade_tooltip}
\alias{trade_tooltip}
\title{Generating tooltip (hover over) text for plots}
\usage{
trade_tooltip(x)
}
\arguments{
\item{x}{data.frame of trade results}
}
\value{
string of HTML formated values formated from the specific results row
to be included in the hover over of a plot
}
\description{
{
trade_tooltip takes a results data.frame from a study and selects the row that
you are hoveringformats numerics to currency for out put in shiny app
}
}
\examples{
trade_tooltip(results)

}

