% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_day_week.R
\name{first_day_week}
\alias{first_day_week}
\title{find first day of week}
\usage{
first_day_week(start_yr, end_yr)
}
\arguments{
\item{start_yr}{Numeric year to begin third friday  calendar}

\item{end_yr}{Numeric year to end third friday calendar}
}
\value{
data.frame \code{monthly_exp_date} with the third friday of each month
when options expire.
}
\description{
{
first_day_week creates a data.frame with the first day of the week (mondays)
to be used in the tidy data process.  The first day of the week is used as a
trade entry criteria in options.shiny and options.studies.
}
}
\examples{
open_weekly <- first_day_week(2010, 2020)
}

